Refactors by pulling functions into small files
