+++ src/js/update.js
@@ -1,0 +1,4 @@
+export function updateResult(grade) {
+  const resultElement = document.querySelector(".grade");
+  resultElement.textContent = grade;
+}
